Add cache updates and optimistic responses on a lot of effects
