+++ typescript/web/src/components/labeling-tool/openlayers-map/edit-label-class.tsx
@@ -154,8 +154,6 @@ export const EditLabelClass = forwardRef<
               name,
               color: newClassColor,
               datasetId,
-              datasetSlug,
-              workspaceSlug,
               imageId,
               previouslySelectedLabelClassId: selectedLabelClassId,
               geometry,
