+++ typescript/web/src/connectors/undo-store/effects/create-label-class.ts
@@ -1,15 +1,19 @@
 import { gql, ApolloClient } from "@apollo/client";
 
+import { v4 as uuid } from "uuid";
 import { useLabelingStore } from "../../labeling-state";
 
 import { Effect } from "..";
-import { getDatasetsQuery } from "../../../pages/[workspaceSlug]/datasets";
-import { datasetLabelClassesQuery } from "../../../components/dataset-class-list/class-item";
+
+import { createLabelClassMutationUpdate } from "./cache-updates/create-label-class-mutation-update";
+import { deleteLabelClassMutationUpdate } from "./cache-updates/delete-label-class-mutation-update";
 
 const createLabelClassQuery = gql`
   mutation createLabelClass($data: LabelClassCreateInput!) {
     createLabelClass(data: $data) {
       id
+      name
+      color
     }
   }
 `;
@@ -27,15 +31,11 @@ export const createCreateLabelClassEffect = (
     name,
     color,
     datasetId,
-    datasetSlug,
-    workspaceSlug,
     selectedLabelClassIdPrevious,
   }: {
     name: string;
     color: string;
     datasetId: string;
-    datasetSlug: string;
-    workspaceSlug: string;
     selectedLabelClassIdPrevious: string | null;
   },
   {
@@ -45,25 +45,23 @@ export const createCreateLabelClassEffect = (
   }
 ): Effect => ({
   do: async () => {
-    const {
-      data: {
-        createLabelClass: { id: labelClassId },
-      },
-    } = await client.mutate({
+    const labelClassId = uuid();
+
+    await client.mutate({
       mutation: createLabelClassQuery,
-      variables: { data: { name, color, datasetId } },
-      refetchQueries: [
-        "getLabelClassesOfDataset",
-        { query: getDatasetsQuery },
-        {
-          query: datasetLabelClassesQuery,
-          variables: { slug: datasetSlug, workspaceSlug },
+      variables: { data: { name, color, datasetId, id: labelClassId } },
+      update: createLabelClassMutationUpdate(datasetId),
+      optimisticResponse: {
+        createLabelClass: {
+          id: labelClassId,
+          name,
+          color,
+          __typename: "LabelClass",
         },
-      ],
+      },
     });
 
     useLabelingStore.setState({ selectedLabelClassId: labelClassId });
-
     return labelClassId;
   },
   undo: async (labelClassId: string) => {
@@ -72,14 +70,10 @@ export const createCreateLabelClassEffect = (
       variables: {
         where: { id: labelClassId },
       },
-      refetchQueries: [
-        "getLabelClassesOfDataset",
-        { query: getDatasetsQuery },
-        {
-          query: datasetLabelClassesQuery,
-          variables: { slug: datasetSlug, workspaceSlug },
-        },
-      ],
+      update: deleteLabelClassMutationUpdate(datasetId),
+      optimisticResponse: {
+        deleteLabelClass: { __typename: "LabelClass", id: labelClassId },
+      },
     });
 
     useLabelingStore.setState({
@@ -92,14 +86,15 @@ export const createCreateLabelClassEffect = (
     await client.mutate({
       mutation: createLabelClassQuery,
       variables: { data: { name, color, id: labelClassId, datasetId } },
-      refetchQueries: [
-        "getLabelClassesOfDataset",
-        { query: getDatasetsQuery },
-        {
-          query: datasetLabelClassesQuery,
-          variables: { slug: datasetSlug, workspaceSlug },
+      update: createLabelClassMutationUpdate(datasetId),
+      optimisticResponse: {
+        createLabelClass: {
+          id: labelClassId,
+          name,
+          color,
+          __typename: "LabelClass",
         },
-      ],
+      },
     });
 
     useLabelingStore.setState({ selectedLabelClassId: labelClassId });
